+++ ffs/api/api_jobs.go
@@ -1,0 +1,57 @@
+package api
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/textileio/powergate/ffs"
+	"github.com/textileio/powergate/ffs/scheduler"
+)
+
+// WatchJobs subscribes to Job status changes. If jids is empty, it subscribes to
+// all Job status changes corresonding to the instance. If jids is not empty,
+// it immediately sends current state of those Jobs. If empty, it doesn't.
+func (i *API) WatchJobs(ctx context.Context, c chan<- ffs.Job, jids ...ffs.JobID) error {
+	var jobs []ffs.Job
+	for _, jid := range jids {
+		j, err := i.sched.GetJob(jid)
+		if err == scheduler.ErrNotFound {
+			continue
+		}
+		if err != nil {
+			return fmt.Errorf("getting current job state: %s", err)
+		}
+		jobs = append(jobs, j)
+	}
+
+	ch := make(chan ffs.Job, 1)
+	for _, j := range jobs {
+		select {
+		case ch <- j:
+		default:
+			log.Warnf("dropped notifying current job state on slow receiver on %s", i.cfg.ID)
+		}
+	}
+	var err error
+	go func() {
+		err = i.sched.WatchJobs(ctx, ch, i.iid)
+		close(ch)
+	}()
+	for j := range ch {
+		if len(jids) == 0 {
+			c <- j
+		}
+	JidLoop:
+		for _, jid := range jids {
+			if jid == j.ID {
+				c <- j
+				break JidLoop
+			}
+		}
+	}
+	if err != nil {
+		return fmt.Errorf("scheduler listener: %s", err)
+	}
+
+	return nil
+}
