+++ ffs/api/option.go
@@ -54,3 +54,39 @@ func (pc PushConfig) Validate() error {
 	}
 	return nil
 }
+
+// NewAddressOption is a function that changes a NewAddressConfig
+type NewAddressOption func(config *NewAddressConfig)
+
+// WithMakeDefault specifies if the new address should become the default
+func WithMakeDefault(makeDefault bool) NewAddressOption {
+	return func(c *NewAddressConfig) {
+		c.makeDefault = makeDefault
+	}
+}
+
+// WithAddressType specifies the type of address to create
+func WithAddressType(addressType string) NewAddressOption {
+	return func(c *NewAddressConfig) {
+		c.addressType = addressType
+	}
+}
+
+// GetLogsOption is a function that changes GetLogsConfig.
+type GetLogsOption func(config *GetLogsConfig)
+
+// WithJidFilter filters only log messages of a Cid related to
+// the Job with id jid.
+func WithJidFilter(jid ffs.JobID) GetLogsOption {
+	return func(c *GetLogsConfig) {
+		c.jid = jid
+	}
+}
+
+// WithHistory indicates that prior log history should be sent
+// to the channel before getting realtime logs.
+func WithHistory(enabled bool) GetLogsOption {
+	return func(c *GetLogsConfig) {
+		c.history = enabled
+	}
+}
