+++ ffs/api/types.go
@@ -57,45 +57,9 @@ type NewAddressConfig struct {
 	addressType string
 }
 
-// NewAddressOption is a function that changes a NewAddressConfig
-type NewAddressOption func(config *NewAddressConfig)
-
-// WithMakeDefault specifies if the new address should become the default
-func WithMakeDefault(makeDefault bool) NewAddressOption {
-	return func(c *NewAddressConfig) {
-		c.makeDefault = makeDefault
-	}
-}
-
-// WithAddressType specifies the type of address to create
-func WithAddressType(addressType string) NewAddressOption {
-	return func(c *NewAddressConfig) {
-		c.addressType = addressType
-	}
-}
-
 // GetLogsConfig contains configuration for a stream-log
 // of human-friendly messages for a Cid execution.
 type GetLogsConfig struct {
 	jid     ffs.JobID
 	history bool
-}
-
-// GetLogsOption is a function that changes GetLogsConfig.
-type GetLogsOption func(config *GetLogsConfig)
-
-// WithJidFilter filters only log messages of a Cid related to
-// the Job with id jid.
-func WithJidFilter(jid ffs.JobID) GetLogsOption {
-	return func(c *GetLogsConfig) {
-		c.jid = jid
-	}
-}
-
-// WithHistory indicates that prior log history should be sent
-// to the channel before getting realtime logs.
-func WithHistory(enabled bool) GetLogsOption {
-	return func(c *GetLogsConfig) {
-		c.history = enabled
-	}
 }
